5 dockerize (#3774)
* Create Docker base image

* Add rough (but failing) test infrastructure

* some tests fail because they expect the original socorro file structure
* some Elasticsearch-using tests are failing, but I'm not sure why because some
  pass fine

Out of 761 tests, there are ~67 errors/failures.

* Add resource.fs.fs_root environment variable to test service

This gets all the tetsts working.

* Move secrets and environment-related config to devbase service

* More scaffolding for the processor

* WIP webapp

* Fix logging in startup so it tell us errors

At startup, there is no self.config.logger, so if things go awry, then calling
self.config.logger.critical doesn't work and then you only see that problem and
not the original problem that occurred which is pretty unhelpful.

Is this a great fix? No. But it's better than before.

* Data setup

* add setup for elasticsearch
* add setup for postgres

* Get the webapp, processor, and tests working

This adds a README with some rough documentation for this environment. It's in
the docker/ directory for now, but when things have settled, we should move it
to the docs.

* Fixes from self-review

* Changes from comments

* nix entrypoint for command which simplifies things
* make some more environment variables
* add some comments and a header

* Fix set_up_stackwalk.sh and clean up shell scripts

If it's a linux 64 system, then we want to use the pre-built breakpad binary to
build stackwalk with. This adds that code.

Also changed some indentation.

Also ran shellcheck and fixed reported issues.